- Added guard logic to AsyncImageView for care retain cycle. - Changed type for rating in WishObject. - Changed UI in the Main.storyboard. - Changed type for listImageView in the ListCell. - Added customisations for listImageView in the ListCellViewModel.

+++ MyWishesRating/MyWishesRating/Modules/List/Common/Data objects/WishObject/WishObject+Decodable.swift
@@ -38,6 +38,6 @@ extension WishObject: Decodable {
             shop = value
         }
         
-        rating = Int.random(in: 1 ..< 10)
+        rating = Float.random(in: 1 ..< 10)
     }
 }
